Refactor to Texture2D *not working* Signed-off-by: Ignacio Quintero <nacho950@gmail.com>

+++ DX11Test/PersonDetector.cpp
@@ -48,8 +48,8 @@ void PersonDetector::captureFrame()
 }
 
 
-vector<float> PersonDetector::getHumans(){
-	vector<float> fringe;
+vector<int> PersonDetector::getHumans(){
+	vector<int> fringe;
 	
 	int count = 0;
 	int color = 0xFF;
@@ -66,13 +66,13 @@ vector<float> PersonDetector::getHumans(){
 
 				if(count == 1 || count & 2 == 1){
 
-					fringe.push_back((float)j);
+					fringe.push_back(j);
 				
 				}
 			}
 		}
 		if(count == 0){
-			fringe.push_back(0.0f);
+			fringe.push_back(0);
 		
 		}
